Corrected LinkedDataIdentical test
    * now thoroughly checks the  `reflect.ValueOf` a Datapoint's `data interface{}` field, and properly handles both pointer and
 non-pointer concrete types in the interface, depending on the `Kind` of `reflect.Value` returned by the `Data()` method on `Dat
apoint` *(i.e. whether the concrete `data interface{}`  is of `type T` or `type *T`)*, and does a `reflect.Indirect` if necessar
y to compare the underlying *value* returned by the `Data()` method.

    * added additional values in the `interfaces` testing slice from a pointer to an object to the actual object a map type to t
he map itself to broadly test the above, and added matching `Datapoints` in the `constructorInputs` slice to match.